fix: rework on useHook
